Move storage protocols to application
Improves direction of imports and dependencies. This puts the "ports" in
the application layer, while keeping the "adapter" in io.
