+++ src/facts_experiment_builder/io/module_registry.py
@@ -2,23 +2,10 @@ from pathlib import Path
 import subprocess
 from pydantic import ValidationError
 import yaml
-from typing import Protocol
 
 # ---------------------- Core imports ----------------------------
 from facts_experiment_builder.core.module.module_schema import ModuleSchema
 from facts_experiment_builder.io.exceptions import ModuleYamlNotFoundError
-
-
-# This is the port
-class ModuleRegistry(Protocol):
-    """Port where module definitions come from.
-
-    application code depends on this class.
-    """
-
-    def get_schema(self, module_name: str) -> ModuleSchema: ...
-    def module_names(self) -> frozenset[str]: ...
-    def version(self) -> str: ...
 
 
 class ModuleRegistryNotFound(Exception):
